+++ backend/canisters/group/impl/src/lifecycle/mod.rs
@@ -1,4 +1,4 @@
-use crate::{init_state as set_state, regular_jobs, Data, RuntimeState, LOG_MESSAGES, WASM_VERSION};
+use crate::{regular_jobs, Data, RuntimeState, WASM_VERSION};
 use types::{Timestamped, Version};
 use utils::env::Environment;
 
@@ -9,17 +9,11 @@ mod pre_upgrade;
 
 const UPGRADE_BUFFER_SIZE: usize = 1024 * 1024; // 1MB
 
-fn init_logger(enable_trace: bool) {
-    let log_messages = canister_logger::init_logger(enable_trace, None, ic_cdk::api::time);
-
-    LOG_MESSAGES.with(|c| *c.borrow_mut() = log_messages);
-}
-
 fn init_state(env: Box<dyn Environment>, data: Data, wasm_version: Version) {
     let now = env.now();
     let regular_jobs = regular_jobs::build();
     let runtime_state = RuntimeState::new(env, data, regular_jobs);
 
-    set_state(runtime_state);
+    crate::init_state(runtime_state);
     WASM_VERSION.with(|v| *v.borrow_mut() = Timestamped::new(wasm_version, now));
 }