+++ backend/canisters/group/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Add SNS1 token to backend ([#2975](https://github.com/open-ic/open-chat/pull/2975))
+- Add ckBTC token to backend ([#2981](https://github.com/open-ic/open-chat/pull/2981))
+
 ### Changed
 
 - Skip processing notifications with no recipients ([#2979](https://github.com/open-ic/open-chat/pull/2979))